requirements.txt
absl-py==0.7.0
appnope==0.1.0
astor==0.7.1
atari-py==0.1.7
autopep8==1.4.3
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.2
entrypoints==0.3
flake8==3.7.7
future==0.17.1
gast==0.2.2
grpcio==1.18.0
gym==0.11.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.2.0
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
mccabe==0.6.1
mock==2.0.0
moviepy==1.0.0
numpy==1.16.1
nvidia-ml-py3==7.352.0
opencv-python==4.0.0.21
parso==0.3.4
pbr==5.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
proglog==0.1.9
prompt-toolkit==2.0.9
protobuf==3.6.1
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.0
pyglet==1.3.2
Pygments==2.3.1
PyOpenGL==3.1.0
PyYAML==3.13
requests==2.21.0
scipy==1.2.1
six==1.12.0
tensorboard==1.12.2
tensorboardX==1.6
tensorflow==1.13.0rc2
tensorflow-estimator==1.13.0
termcolor==1.1.0
torch==1.0.1.post2
torchvision==0.2.1
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
